// src/proto/types/fundingRounds/GetOneResponse.ts
// Original file: src/proto/proto/fundingRound.proto

import type { FundingRound as _fundingRounds_FundingRound, FundingRound__Output as _fundingRounds_FundingRound__Output } from '../fundingRounds/FundingRound';

export interface GetOneResponse {
  'message'?: (string);
  'data'?: (_fundingRounds_FundingRound | null);
  '_data'?: "data";
}

export interface GetOneResponse__Output {
  'message': (string);
  'data'?: (_fundingRounds_FundingRound__Output | null);
  '_data': "data";
}