// src/proto/types/fundingRounds/QueryFundingRound.ts
// Original file: src/proto/proto/fundingRound.proto


export interface QueryFundingRound {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'page'?: (number);
  'limit'?: (number);
  'sortBy'?: (string);
  'sortOrder'?: (string);
  'investeeId'?: (number);
  'announcedDate'?: (string);
  'round'?: (string);
  'type'?: (string);
  'preMoney'?: (number | string);
  'moneyRaised'?: (number | string);
  'description'?: (string);
  '_id'?: "id";
  '_createdAt'?: "createdAt";
  '_updatedAt'?: "updatedAt";
  '_createdBy'?: "createdBy";
  '_updatedBy'?: "updatedBy";
  '_state'?: "state";
  '_page'?: "page";
  '_limit'?: "limit";
  '_sortBy'?: "sortBy";
  '_sortOrder'?: "sortOrder";
  '_investeeId'?: "investeeId";
  '_announcedDate'?: "announcedDate";
  '_round'?: "round";
  '_type'?: "type";
  '_preMoney'?: "preMoney";
  '_moneyRaised'?: "moneyRaised";
  '_description'?: "description";
}

export interface QueryFundingRound__Output {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'page'?: (number);
  'limit'?: (number);
  'sortBy'?: (string);
  'sortOrder'?: (string);
  'investeeId'?: (number);
  'announcedDate'?: (string);
  'round'?: (string);
  'type'?: (string);
  'preMoney'?: (number);
  'moneyRaised'?: (number);
  'description'?: (string);
  '_id': "id";
  '_createdAt': "createdAt";
  '_updatedAt': "updatedAt";
  '_createdBy': "createdBy";
  '_updatedBy': "updatedBy";
  '_state': "state";
  '_page': "page";
  '_limit': "limit";
  '_sortBy': "sortBy";
  '_sortOrder': "sortOrder";
  '_investeeId': "investeeId";
  '_announcedDate': "announcedDate";
  '_round': "round";
  '_type': "type";
  '_preMoney': "preMoney";
  '_moneyRaised': "moneyRaised";
  '_description': "description";
}